+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Fire control".
 - Tests for "Genetics and computers - part 1".
 - Tests for "Retro typewriter art".
+- Tests for "Parse SQL queries".
 
 ### Changed
 - Renaming "Linear Bézier curves" to "Cubic Bézier curves".